chore: split srv-driven cadl ranch directory
