--- packages/http-client-csharp/generator/TestProjects/CadlRanch/http/resiliency/srv-driven/Configuration.json
@@ -1,6 +1,0 @@
-{
-  "output-folder": ".",
-  "namespace": "Resiliency.ServiceDriven",
-  "library-name": "Resiliency.ServiceDriven",
-  "use-model-reader-writer": true
-}
