+++ packages/http-client-csharp/generator/TestProjects/CadlRanch/http/resiliency/srv-driven/v1/Configuration.json
@@ -1,6 +1,6 @@
 {
   "output-folder": ".",
-  "namespace": "Resiliency.ServiceDriven.V1",
-  "library-name": "Resiliency.ServiceDriven.V1",
+  "namespace": "Resiliency.SrvDriven.V1",
+  "library-name": "Resiliency.SrvDriven.V1",
   "use-model-reader-writer": true
 }
