+++ packages/http-client-csharp/generator/TestProjects/CadlRanch/http/resiliency/srv-driven/v1/src/Resiliency.SrvDriven.V1.csproj
@@ -1,9 +1,9 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <Description>This is the Resiliency.ServiceDriven.V1 client library for developing .NET applications with rich experience.</Description>
-    <AssemblyTitle>SDK Code Generation Resiliency.ServiceDriven.V1</AssemblyTitle>
+    <Description>This is the Resiliency.SrvDriven.V1 client library for developing .NET applications with rich experience.</Description>
+    <AssemblyTitle>SDK Code Generation Resiliency.SrvDriven.V1</AssemblyTitle>
     <Version>1.0.0-beta.1</Version>
-    <PackageTags>Resiliency.ServiceDriven.V1</PackageTags>
+    <PackageTags>Resiliency.SrvDriven.V1</PackageTags>
     <TargetFramework>netstandard2.0</TargetFramework>
     <LangVersion>latest</LangVersion>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
