+++ packages/http-client-csharp/generator/TestProjects/CadlRanch/http/resiliency/srv-driven/v2/Configuration.json
@@ -1,0 +1,6 @@
+{
+  "output-folder": ".",
+  "namespace": "Resiliency.SrvDriven.V2",
+  "library-name": "Resiliency.SrvDriven.V2",
+  "use-model-reader-writer": true
+}
