+++ packages/http-client-csharp/generator/TestProjects/CadlRanch/http/resiliency/srv-driven/v2/tspCodeModel.json
